refactor: promote the routing-status store to a component and make its tests pin what they name
Tier-3 review scored elegance 6/10 against the required 7. One file held four
cohesive facets across 2,051 lines — transaction SQL, artifact observation and
publication, quarantine maintenance, and durable identity with its DDL — and the
quarantine operation alone was 286 lines at complexity 56. Each facet is now its
own file behind an index that publishes the same surface, so no importer changes
beyond the path. The move is pure: no declaration was rewritten and no test
expectation changed.

The quarantine protocol threaded ten parameters plus mutation state, two effect
ledgers, paths and an ownership assertion through separate frames, and rebuilt
the same ownership rule again in clear. One maintenance context owns them, the
phases are named, and both paths share the rule. The CLI rendered the same six
effect fields twice, where a new field could have diverged between copies.

The tests were scored 5/10 and had three defects worth the name. Two assertions
matched source text, so a correct refactor would have failed them while the value
assertion beside them already pinned the property. Fault injection fired on the
Nth call rather than the named event, which had already moved a fault to the
wrong state twice on this branch; triggers are now the events themselves. A
simulation test opened a read-only handle and wrote through it, pinning behaviour
only the fake exhibits.

One test claimed to cover the ownership recheck after the main database moves and
did not: deleting that check left all 95 tests passing. Its trigger fired on the
quarantine-root sync, which an earlier check caught. It now loses ownership after
the source unlink and its barrier, and fails when the check is removed.

The layering invariant's forbidden-target set still named the pre-move filename,
so it had stopped protecting; it now matches the component by prefix, verified by
injecting the import it must reject.

Identity unchanged: 2b1df746…55d9 / 1723384134.
