+++ src/coordinator/handoff-routing/runner.ts
@@ -26,7 +26,7 @@ import type { TimePort, TimerHandle } from '../../infra/port-types.js';
 import type { RecordedProcessIdentity } from '../../infra/process-containment.js';
 import type { Runtime } from '../../runtime/ports.js';
 import { createRealRuntime } from '../../runtime/real.js';
-import { handoffRoutingStatusGeneration } from '../../store/handoff-routing-status-store.js';
+import { handoffRoutingStatusGeneration } from '../../store/handoff-routing-status-store/index.js';
 import { createIpcClient } from '../../transport/ipc/client.js';
 import {
   HANDOFF_ROUTING_BASIS_OBLIGATIONS,
